Move living-assets-evolution pallet to frontier
