+++ evolution-chain/runtime/src/lib.rs
@@ -46,8 +46,6 @@ use pallet_transaction_payment::{ConstFeeMultiplier, CurrencyAdapter, Multiplier
 pub use sp_runtime::BuildStorage;
 pub use sp_runtime::{Perbill, Permill};
 
-pub use pallet_living_assets_evolution;
-
 /// An index to a block.
 pub type BlockNumber = evochain_primitives::BlockNumber;
 
@@ -271,18 +269,6 @@ impl pallet_sudo::Config for Runtime {
 	type WeightInfo = pallet_sudo::weights::SubstrateWeight<Runtime>;
 }
 
-parameter_types! {
-	/// Max length of the `TokenUri`
-	pub const MaxTokenUriLength: u32 = 512;
-}
-
-/// Configure the pallet-living-assets-evolution in pallets/living-assets-evolution.
-impl pallet_living_assets_evolution::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type WeightInfo = pallet_living_assets_evolution::weights::SubstrateWeight<Runtime>;
-	type MaxTokenUriLength = MaxTokenUriLength;
-}
-
 impl pallet_bridge_grandpa::Config for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	type BridgedChain = bp_rococo::Rococo;
@@ -301,9 +287,6 @@ construct_runtime!(
 		Balances: pallet_balances,
 		TransactionPayment: pallet_transaction_payment,
 		Sudo: pallet_sudo,
-
-		// Include the custom logic from the pallet-living-assets-evolution in the runtime.
-		LivingAssetsEvolution: pallet_living_assets_evolution,
 
 		BridgeRococoGrandpa: pallet_bridge_grandpa,
 	}
@@ -352,7 +335,6 @@ mod benches {
 		[frame_system, SystemBench::<Runtime>]
 		[pallet_balances, Balances]
 		[pallet_timestamp, Timestamp]
-		[pallet_living_assets_evolution, LivingAssetsEvolution]
 	);
 }
