+++ frontier-chain/runtime/src/lib.rs
@@ -364,6 +364,18 @@ impl pallet_hotfix_sufficients::Config for Runtime {
 	type WeightInfo = pallet_hotfix_sufficients::weights::SubstrateWeight<Self>;
 }
 
+parameter_types! {
+	/// Max length of the `TokenUri`
+	pub const MaxTokenUriLength: u32 = 512;
+}
+
+/// Configure the pallet-living-assets-evolution in pallets/living-assets-evolution.
+impl pallet_living_assets_evolution::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+	type WeightInfo = pallet_living_assets_evolution::weights::SubstrateWeight<Self>;
+	type MaxTokenUriLength = MaxTokenUriLength;
+}
+
 // Create the runtime by composing the FRAME pallets that were previously configured.
 construct_runtime!(
 	pub enum Runtime {
@@ -380,6 +392,9 @@ construct_runtime!(
 		DynamicFee: pallet_dynamic_fee,
 		BaseFee: pallet_base_fee,
 		HotfixSufficients: pallet_hotfix_sufficients,
+
+		// Local pallets
+		LivingAssetsEvolution: pallet_living_assets_evolution,
 	}
 );
 
@@ -511,6 +526,7 @@ mod benches {
 		[pallet_timestamp, Timestamp]
 		[pallet_sudo, Sudo]
 		[pallet_evm, EVM]
+		[pallet_living_assets_evolution, LivingAssetsEvolution]
 	);
 }
